requirements.txt
discord.py==2.7.1
httpx==0.28.1
python-dotenv==1.2.1
apscheduler==3.11.2
aiohttp==3.8.6
